+++ scanner/Options.py
@@ -132,9 +132,10 @@ config['browser_unsupported_mime_types'] = ['image/tiff', 'image/webp', 'image/x
 # json_version = 4.22 since albumCacheBase removed from mediaList elements
 # json_version = 4.23 since subalbums have complete path
 # json_version = 4.24 since default albums and media sorting implemented
+# json_version = 4.25 since the bugs in reading exif metadata have been corrected
 
 # json_version = 0
-json_version = 4.24
+json_version = 4.25
 
 
 def set_obsolete_json_version_flag():